requirement.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-dotenv
pydantic
groq
gspread
oauth2client
python-dotenv
